spriteops_remove_transparent_colour

+++ SpriteOperations/dllmain.cpp
@@ -15,7 +15,7 @@ ex double spriteops_get_cropped_dimensions(int* data, double w, double h, double
 	return 1.0;
 }
 
-ex double spriteops_remove_transparent_colour(float* data, double length, double colour) {
+ex double spriteops_remove_transparent_colour(int* data, double length, double colour) {
 	spriteops::spriteops_remove_transparent_colour(data, BYTES2FLOATS(length), (int)colour);
 	return 1.0;
 }
